requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.44.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
narwhals==1.36.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pefile==2023.2.7
pillow==11.2.1
plotly==6.0.1
preshed==3.0.9
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
shap==0.47.2
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
statsmodels==0.14.4
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
